requirements.txt
asgiref==3.8.1
attrs==24.3.0
Django==5.1.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.28.0
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
PyJWT==2.10.1
python-dotenv==1.0.1
PyYAML==6.0.2
referencing==0.35.1
rpds-py==0.22.3
sqlparse==0.5.3
tzdata==2024.2
uritemplate==4.1.1
